last commit

+++ backend/migrations/20230702113104-create-messages.js
@@ -9,17 +9,17 @@ module.exports = {
         type: DataTypes.INTEGER,
       },
       messageId: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         primaryKey: true,
         autoIncrement: true,
         allowNull: false,
       },
       senderId: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
       recipientId: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
       chatId: {
